# bibliotheque-api/swagger-spring/src/main/java/io/swagger/api/UtilisateurApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.7).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Utilisateur;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-08-05T10:23:06.042Z")

@Api(value = "utilisateur", description = "the utilisateur API")
public interface UtilisateurApi {

    @ApiOperation(value = "Ajouter une nouvelle réservation", nickname = "addUtilisateur", notes = "", tags={ "utilisateur", })
    @ApiResponses(value = { 
        @ApiResponse(code = 405, message = "Invalid input") })
    @RequestMapping(value = "/utilisateur",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Void> addUtilisateur(@ApiParam(value = "Un objet Reservation doit être envoyé pour être ajouté" ,required=true )  @Valid @RequestBody Utilisateur body);


    @ApiOperation(value = "Supprimer un livre", nickname = "deleteUtilisateur", notes = "", tags={ "utilisateur", })
    @ApiResponses(value = { 
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "livre not found") })
    @RequestMapping(value = "/utilisateur/{utilisateurId}",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteUtilisateur(@ApiParam(value = "ID du livre à supprimer",required=true) @PathVariable("utilisateurId") Long utilisateurId,@ApiParam(value = "" ) @RequestHeader(value="api_key", required=false) String apiKey);


    @ApiOperation(value = "Trouver des réservations", nickname = "findUtilisateursByNom", notes = "Plusieurs valeurs peuvent être séparées par une virgule", response = Utilisateur.class, responseContainer = "List", tags={ "utilisateur", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Utilisateur.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid status value") })
    @RequestMapping(value = "/utilisateur",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Utilisateur>> findUtilisateursByNom(@NotNull @ApiParam(value = "Status values that need to be considered for filter", required = true, allowableValues = "available, pending, sold") @Valid @RequestParam(value = "status", required = true) List<String> status);


    @ApiOperation(value = "Trouve une réservation par son ID", nickname = "getUtilisateurById", notes = "Trouve une réservation par son ID", response = Utilisateur.class, tags={ "utilisateur", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Utilisateur.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "livre not found") })
    @RequestMapping(value = "/utilisateur/{utilisateurId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Utilisateur> getUtilisateurById(@ApiParam(value = "ID of livre to return",required=true) @PathVariable("utilisateurId") Long utilisateurId);


    @ApiOperation(value = "Mettre à jour un livre avec un form data", nickname = "updateUtilisateurWithForm", notes = "", tags={ "utilisateur", })
    @ApiResponses(value = { 
        @ApiResponse(code = 405, message = "Invalid input") })
    @RequestMapping(value = "/utilisateur/{utilisateurId}",
        produces = { "application/json" }, 
        consumes = { "application/x-www-form-urlencoded" },
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateUtilisateurWithForm(@ApiParam(value = "ID du livre qui doit être mis à jour",required=true) @PathVariable("utilisateurId") Long utilisateurId,@ApiParam(value = "Mettre à jour le nom du livre") @RequestParam(value="name", required=false)  String name,@ApiParam(value = "Mettre à jour l'auteur d'un livre") @RequestParam(value="auteur", required=false)  String auteur,@ApiParam(value = "Mettre à jour la quantité d'un livre") @RequestParam(value="quantite", required=false)  Integer quantite);

}